app.js共通部分の外だしとfilter部分の修正
