+++ .env
@@ -2,4 +2,4 @@ MYSQL_HOST = localhost
 MYSQL_PORT = 3306
 MYSQL_USERNAME = satoru
 MYSQL_PASSWORD = 3106
-MYSQL_DATABASE = 
+MYSQL_DATABASE = food_protein
